+++ apps/common/lib/lexical/ast/env.ex
@@ -1,17 +1,10 @@
-defmodule Lexical.Server.CodeIntelligence.Completion.Env do
+defmodule Lexical.Ast.Env do
   alias Future.Code, as: Code
   alias Lexical.Ast
-  alias Lexical.Completion.Builder
-  alias Lexical.Completion.Environment
+  alias Lexical.Ast.Environment
   alias Lexical.Document
-  alias Lexical.Document.Edit
   alias Lexical.Document.Position
-  alias Lexical.Document.Range
   alias Lexical.Project
-  alias Lexical.Protocol.Types.Completion
-  alias Lexical.Server.CodeIntelligence.Completion.Env
-
-  import Document.Line
 
   defstruct [
     :project,
@@ -72,7 +65,21 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
   end
 
   @impl Environment
-  def in_context?(%__MODULE__{} = env, :function_capture) do
+  def in_context?(%__MODULE__{} = env, context_type) do
+    do_in_context?(env, context_type)
+  end
+
+  @impl Environment
+  def empty?("") do
+    true
+  end
+
+  @impl Environment
+  def empty?(string) when is_binary(string) do
+    String.trim(string) == ""
+  end
+
+  defp do_in_context?(env, :function_capture) do
     env
     |> prefix_token_stream()
     |> Enum.reduce_while(false, fn
@@ -95,8 +102,7 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end)
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :struct_reference) do
+  defp do_in_context?(env, :struct_reference) do
     case cursor_context(env) do
       {:ok, _line, {:struct, _}} ->
         true
@@ -120,8 +126,7 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :struct_arguments) do
+  defp do_in_context?(env, :struct_arguments) do
     cursor_path = Ast.cursor_path(env.document, env.position)
 
     Enum.any?(cursor_path, fn
@@ -135,9 +140,8 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end)
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :struct_field_value) do
-    if in_context?(env, :struct_arguments) do
+  defp do_in_context?(env, :struct_field_value) do
+    if do_in_context?(env, :struct_arguments) do
       env
       |> prefix_tokens(2)
       |> Enum.any?(fn
@@ -149,8 +153,7 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :pipe) do
+  defp do_in_context?(env, :pipe) do
     env
     |> prefix_token_stream()
     |> Enum.reduce_while(false, fn
@@ -171,8 +174,7 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end)
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :bitstring) do
+  defp do_in_context?(env, :bitstring) do
     env
     |> prefix_tokens(:all)
     |> Enum.reduce_while(
@@ -185,8 +187,7 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     )
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :alias) do
+  defp do_in_context?(env, :alias) do
     # Aliases are complicated, especially if we're trying to find out if we're in
     # them from the current cursor position moving backwards.
     # I'll try to describe the state machine below.
@@ -231,143 +232,16 @@ defmodule Lexical.Server.CodeIntelligence.Completion.Env do
     end)
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :import) do
+  defp do_in_context?(env, :import) do
     in_directive?(env, ~c"import")
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :use) do
+  defp do_in_context?(env, :use) do
     in_directive?(env, ~c"use")
   end
 
-  @impl Environment
-  def in_context?(%__MODULE__{} = env, :require) do
+  defp do_in_context?(env, :require) do
     in_directive?(env, ~c"require")
-  end
-
-  @impl Environment
-  def empty?("") do
-    true
-  end
-
-  def empty?(string) when is_binary(string) do
-    String.trim(string) == ""
-  end
-
-  @behaviour Builder
-
-  @impl Builder
-  def snippet(%Env{}, snippet_text, options \\ []) do
-    options
-    |> Keyword.put(:insert_text, snippet_text)
-    |> Keyword.put(:insert_text_format, :snippet)
-    |> Completion.Item.new()
-    |> boost(0)
-  end
-
-  @impl Builder
-  def plain_text(%Env{}, insert_text, options \\ []) do
-    options
-    |> Keyword.put(:insert_text, insert_text)
-    |> Completion.Item.new()
-    |> boost(0)
-  end
-
-  @impl Builder
-  def text_edit(%Env{} = env, text, {start_char, end_char}, options \\ []) do
-    line_number = env.position.line
-    range = Range.new(Position.new(line_number, start_char), Position.new(line_number, end_char))
-    edits = Document.Changes.new(env.document, Edit.new(text, range))
-
-    options
-    |> Keyword.put(:text_edit, edits)
-    |> Completion.Item.new()
-    |> boost(0)
-  end
-
-  @impl Builder
-  def text_edit_snippet(%Env{} = env, text, {start_char, end_char}, options \\ []) do
-    line_number = env.position.line
-    range = Range.new(Position.new(line_number, start_char), Position.new(line_number, end_char))
-    edits = Document.Changes.new(env.document, Edit.new(text, range))
-
-    options
-    |> Keyword.put(:text_edit, edits)
-    |> Keyword.put(:insert_text_format, :snippet)
-    |> Completion.Item.new()
-    |> boost(0)
-  end
-
-  @impl Builder
-  def fallback(nil, fallback), do: fallback
-  def fallback("", fallback), do: fallback
-  def fallback(detail, _), do: detail
-
-  @impl Builder
-  def boost(item, local_boost \\ 1, global_boost \\ 0)
-
-  def boost(%Completion.Item{} = item, local_boost, global_boost)
-      when local_boost in 0..9 and global_boost in 0..9 do
-    global_boost = Integer.to_string(9 - global_boost)
-    local_boost = Integer.to_string(9 - local_boost)
-
-    sort_text = "0#{global_boost}#{local_boost}_#{item.label}"
-    %Completion.Item{item | sort_text: sort_text}
-  end
-
-  # end builder behaviour
-
-  @spec strip_struct_reference(t()) :: {Document.t(), Position.t()}
-  def strip_struct_reference(%__MODULE__{} = env) do
-    if in_context?(env, :struct_reference) do
-      do_strip_struct_reference(env)
-    else
-      {env.document, env.position}
-    end
-  end
-
-  # private
-
-  defp do_strip_struct_reference(%__MODULE__{} = env) do
-    completion_length =
-      case Code.Fragment.cursor_context(env.prefix) do
-        {:struct, {:dot, {:alias, struct_name}, []}} ->
-          # add one because of the trailing period
-          length(struct_name) + 1
-
-        {:struct, {:local_or_var, local_name}} ->
-          length(local_name)
-
-        {:struct, struct_name} ->
-          length(struct_name)
-
-        {:local_or_var, local_name} ->
-          length(local_name)
-      end
-
-    column = env.position.character
-    percent_position = column - (completion_length + 1)
-
-    new_line_start = String.slice(env.line, 0, percent_position - 1)
-    new_line_end = String.slice(env.line, percent_position..-1)
-    new_line = [new_line_start, new_line_end]
-    new_position = Position.new(env.position.line, env.position.character - 1)
-    line_to_replace = env.position.line
-
-    new_document =
-      env.document.lines
-      |> Enum.with_index(1)
-      |> Enum.reduce([], fn
-        {line(ending: ending), ^line_to_replace}, acc ->
-          [acc, new_line, ending]
-
-        {line(text: line_text, ending: ending), _}, acc ->
-          [acc, line_text, ending]
-      end)
-      |> IO.iodata_to_binary()
-
-    {new_document, new_position}
   end
 
   defp in_directive?(%__MODULE__{} = env, context_name) do
