+++ apps/server/lib/lexical/server/code_intelligence/completion/translations/callback.ex
@@ -1,6 +1,6 @@
 defmodule Lexical.Server.CodeIntelligence.Completion.Translations.Callback do
+  alias Lexical.Ast.Env
   alias Lexical.RemoteControl.Completion.Candidate
-  alias Lexical.Server.CodeIntelligence.Completion.Env
   alias Lexical.Server.CodeIntelligence.Completion.Translatable
   alias Lexical.Server.CodeIntelligence.Completion.Translations.Callable
