+++ apps/server/lib/lexical/server/code_intelligence/completion/translations/map_field.ex
@@ -1,6 +1,6 @@
 defmodule Lexical.Server.CodeIntelligence.Completion.Translations.MapField do
+  alias Lexical.Ast.Env
   alias Lexical.RemoteControl.Completion.Candidate
-  alias Lexical.Server.CodeIntelligence.Completion.Env
   alias Lexical.Server.CodeIntelligence.Completion.Translatable
 
   use Translatable.Impl, for: Candidate.MapField
