+++ apps/server/lib/lexical/server/code_intelligence/completion/translations/struct_field.ex
@@ -1,7 +1,7 @@
 defmodule Lexical.Server.CodeIntelligence.Completion.Translations.StructField do
   alias Future.Code, as: Code
+  alias Lexical.Ast.Env
   alias Lexical.RemoteControl.Completion.Candidate
-  alias Lexical.Server.CodeIntelligence.Completion.Env
   alias Lexical.Server.CodeIntelligence.Completion.Translatable
 
   use Translatable.Impl, for: Candidate.StructField
