+++ apps/server/test/lexical/server/code_intelligence/completion_test.exs
@@ -1,8 +1,8 @@
 defmodule Lexical.Server.CodeIntelligence.CompletionTest do
+  alias Lexical.Ast.Env
   alias Lexical.Document
   alias Lexical.Protocol.Types.Completion
   alias Lexical.RemoteControl.Completion.Candidate
-  alias Lexical.Server.CodeIntelligence.Completion.Env
 
   use Lexical.Test.Server.CompletionCase
   use Patch
